Clean up ProjectionCanvas.

+++ include/Color.h
@@ -4,8 +4,8 @@
 class Color
 {
 public:
-	float R, G, B;
-	Color(float r, float g, float b) : R(r), G(g), B(b) {};
+	double R, G, B;
+	Color(double r, double g, double b) : R(r), G(g), B(b) {};
 	Color operator+(const Color c)	{ return Color (R+c.R, G+c.G, B+c.B); };
 	Color operator*(const double& b)	{	return Color (R*b, G*b, B*b); };
 	bool operator==(const Color& c)	{	return (c.R==R && c.G==G && c.B == B); };
